+++ frontend/hr-master/src/components/Login/LoginAside/LoginAside.jsx
@@ -1,6 +1,6 @@
 import styles from './LoginAside.module.css';
 
-export function LoginAside() {
+export default function LoginAside() {
 	return (
 		<div className={styles.leftContent}>
 			<div className={styles.textBox}>
